* tif图片预览，根据反馈的意见，已经全部修改 1、配置文件、配置项默认值已改回image、tif 2、去掉了静态工具类上的@component 注解 3、修改OnlinePreview处理，在切换image方式预览时url中加入参数previewType=image，加入对应的处理，兼容在pdf预览模式下切换到JPG方式预览

+++ server/src/main/java/cn/keking/service/impl/TiffFilePreviewImpl.java
@@ -41,6 +41,7 @@ public class TiffFilePreviewImpl implements FilePreview {
     public String filePreviewHandle(String url, Model model, FileAttribute fileAttribute) {
 
         String tifPreviewType = ConfigConstants.getTifPreviewType();
+        String priveiwType = String.valueOf(model.getAttribute("previewType"));
 
         if("tif".equalsIgnoreCase(tifPreviewType)){
 
@@ -71,15 +72,14 @@ public class TiffFilePreviewImpl implements FilePreview {
 
             if(fileJpg.exists()){
                 // 转换后的tif没用了，可以删掉了
-                fileTiff.delete();
+                if(fileTiff.exists()){
+                    fileTiff.delete();
+                }
 
                 String baseUrl = BaseUrlFilter.getBaseUrl();
-                if("pdf".equalsIgnoreCase(tifPreviewType)){
+                if("pdf".equalsIgnoreCase(tifPreviewType) && !"image".equalsIgnoreCase(priveiwType)){
                     File filePdf = ConvertPicUtil.convertJpg2Pdf(fileDir + uuid + ".jpg", fileDir + uuid + ".pdf");
                     if(filePdf.exists()){
-                        // 转换后的jpg没用了，可以删掉了
-                        fileJpg.delete();
-
                         String pdfUrl = baseUrl + uuid + ".pdf";
                         model.addAttribute("pdfUrl", pdfUrl);
